Docker_practice/backendAgent/pyproject.toml
[project]
name = "backendagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "openai-agents>=0.7.0",
]
